rename util module to prevent clash with other util module
